# config/NMDAOpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API CORONA FINANCIERO - ADMISION",
    "description": "API Corona Financiero Admision",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [{
    "url": "http://localhost:8080/admision/api",
    "description": "Generated server url"
  }],
...
  "paths": {
...
    "/perfilamiento/rol": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarRoles",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionRolDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Perfilamiento"],
        "operationId": "actualizarRol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarRolRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Perfilamiento"],
        "operationId": "crearRol",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearRolRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/perfilamiento/perfil": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarPerfiles",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionPerfilDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Perfilamiento"],
        "operationId": "modificarPerfil",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificaPerfilDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Perfilamiento"],
        "operationId": "crearPerfil",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPerfilDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/perfilamiento/perfilFuncionario": {
      "put": {
        "tags": ["Perfilamiento"],
        "operationId": "actualizarPerfilesFuncionario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarPerfilFuncionarioRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": ["Perfilamiento"],
        "operationId": "crearPerfilesFuncionario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPerfilFuncionarioRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/cliente/registro": {
      "post": {
        "tags": ["Cliente"],
        "operationId": "registroProspecto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProspectoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroProspectoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/region-comuna/{codigoRegion}/comunas": {
      "get": {
        "tags": ["Region y Comunas"],
        "operationId": "listarComunas",
        "parameters": [{
          "name": "codigoRegion",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComunaDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/perfilamiento/rol/{rolId}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "obtenerRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RolDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Perfilamiento"],
        "operationId": "eliminarRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/perfilamiento/rol/vigente": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarRolesVigentes",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionRolDTO"
                }
              }
            }
          }
        }
      }
    },
    "/perfilamiento/perfil/{perfilId}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "obtenerPerfil",
        "parameters": [{
          "name": "perfilId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Perfilamiento"],
        "operationId": "eliminarPerfil",
        "parameters": [{
          "name": "perfilId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/perfilamiento/perfil/vigente": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarPerfilesVigentesPorFiltro",
        "parameters": [{
          "name": "pagina",
          "in": "query",
          "required": false,
          "schema": {
            "type": "integer",
            "format": "int32",
            "default": 0
          }
        }, {
          "name": "filtro",
          "in": "query",
          "required": false,
          "schema": {
            "type": "string",
            "default": ""
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginacionPerfilResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/perfilamiento/perfil/funcionario/{rutFuncionario}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "listarPerfilesPorFuncionario",
        "parameters": [{
          "name": "rutFuncionario",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PerfilResponseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/cliente/{rutCliente}": {
      "get": {
        "tags": ["Cliente"],
        "operationId": "buscarCliente",
        "parameters": [{
          "name": "rutCliente",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/cliente/datos/{rutCliente}": {
      "get": {
        "tags": ["Cliente"],
        "operationId": "datosPersonalesBasicos",
        "parameters": [{
          "name": "rutCliente",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DatosBasicosMinimosResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualizarRolRequestDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "modulo": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "ModificaPerfilDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "perfilId": {
            "type": "integer",
            "format": "int64"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "ActualizarPerfilFuncionarioRequestDTO": {
        "type": "object",
        "properties": {
          "rutFuncionario": {
            "type": "string"
          },
          "perfiles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "CrearRolRequestDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "modulo": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "CrearPerfilDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "CrearPerfilFuncionarioRequestDTO": {
        "type": "object",
        "properties": {
          "funcionario": {
            "$ref": "#/components/schemas/FuncionarioRequestDTO"
          },
          "perfiles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "FuncionarioRequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "nombreUsuario": {
            "type": "string"
          },
          "sucursalId": {
            "type": "integer",
            "format": "int64"
          },
          "nombreSucursal": {
            "type": "string"
          },
          "cargoId": {
            "type": "integer",
            "format": "int64"
          },
          "nombreCargo": {
            "type": "string"
          }
        }
      },
...
      "ProspectoDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "movil": {
            "type": "string"
          }
        }
      },
      "RegistroProspectoResponseDTO": {
        "type": "object",
        "properties": {
          "exito": {
            "type": "boolean"
          },
          "mensaje": {
            "type": "string"
          }
        }
      },
...
      "ComunaDTO": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          }
        }
      },
...
      "ListarRolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "PaginacionRolDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListarRolDTO"
            }
          }
        }
      },
      "RolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "moduloId": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "ListarPerfilDTO": {
        "type": "object",
        "properties": {
          "perfilId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "PaginacionPerfilDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "perfiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListarPerfilDTO"
            }
          }
        }
      },
      "PerfilDTO": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "perfilId": {
            "type": "integer",
            "format": "int64"
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
      "PaginacionPerfilResponseDTO": {
        "type": "object",
        "properties": {
          "totalPagina": {
            "type": "integer",
            "format": "int64"
          },
          "pagina": {
            "type": "integer",
            "format": "int64"
          },
          "totalElementos": {
            "type": "integer",
            "format": "int64"
          },
          "perfiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PerfilResponseDTO"
            }
          }
        }
      },
      "PerfilResponseDTO": {
        "type": "object",
        "properties": {
          "perfilId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
...
      "ProspectoResponseDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "movil": {
            "type": "string"
          },
          "prospectoId": {
            "type": "integer",
            "format": "int64"
          },
          "solicitudId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DatosBasicosMinimosResponseDTO": {
        "type": "object",
        "properties": {
          "prospectoId": {
            "type": "integer",
            "format": "int64"
          },
          "rut": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "apellidoPatenro": {
            "type": "string"
          },
          "apellidoMaterno": {
            "type": "string"
          }
        }
      },
...
    }
  }
}